fix critical issues
