+++ src/components/Loader/PeoplePage.tsx
@@ -2,26 +2,23 @@ import { useEffect, useState } from 'react';
 import { Person } from '../../types';
 import { Loader } from './Loader';
 import { getPeople } from '../../api';
-import classNames from 'classnames';
-import { useLocation } from 'react-router-dom';
-import { PersonLink } from './PersonLink';
+import { PeopleTable } from './PeopleTable';
 
 export const PeoplePage = () => {
   const [people, setPeople] = useState<Person[]>([]);
   const [isLoading, setIsLoading] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
-  const { pathname } = useLocation();
 
   useEffect(() => {
     setIsLoading(true);
     setErrorMessage('');
 
     getPeople()
-      .then(peopleFromServer => {
-        setPeople(peopleFromServer);
+      .then(peopleFromSerever => {
+        setPeople(peopleFromSerever);
       })
       .catch(() => {
-        setErrorMessage('Something went wrong');
+        setErrorMessage('Something went wrong!');
       })
       .finally(() => {
         setIsLoading(false);
@@ -46,73 +43,7 @@ export const PeoplePage = () => {
             <p data-cy="noPeopleMessage">There are no people on the server</p>
           )}
 
-          {!isLoading && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  <th>Name</th>
-                  <th>Sex</th>
-                  <th>Born</th>
-                  <th>Died</th>
-                  <th>Mother</th>
-                  <th>Father</th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {people.map(person => {
-                  const mother = people.find(p => p.name === person.motherName);
-                  const father = people.find(p => p.name === person.fatherName);
-                  const isSelected = pathname === `/people/${person.slug}`;
-
-                  return (
-                    <tr
-                      key={person.slug}
-                      data-cy="person"
-                      className={classNames({
-                        'has-background-warning': isSelected,
-                      })}
-                    >
-                      <td>
-                        <PersonLink person={person} />
-                      </td>
-
-                      <td>{person.sex}</td>
-                      <td>{person.born}</td>
-                      <td>{person.died}</td>
-
-                      <td>
-                        {person.motherName ? (
-                          mother ? (
-                            <PersonLink person={mother} />
-                          ) : (
-                            person.motherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-
-                      <td>
-                        {person.fatherName ? (
-                          father ? (
-                            <PersonLink person={father} />
-                          ) : (
-                            person.fatherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          )}
+          {!isLoading && people.length > 0 && <PeopleTable people={people} />}
         </div>
       </div>
     </>
