+++ src/components/Loader/PeopleTable.tsx
@@ -1,0 +1,80 @@
+import { useLocation } from 'react-router-dom';
+import classNames from 'classnames';
+import { Person } from '../../types';
+import { PersonLink } from './PersonLink';
+
+type Props = {
+  people: Person[];
+};
+
+export const PeopleTable: React.FC<Props> = ({ people }) => {
+  const { pathname } = useLocation();
+
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {people.map(person => {
+          const mother = people.find(p => p.name === person.motherName);
+          const father = people.find(p => p.name === person.fatherName);
+          const isSelected = pathname === `/people/${person.slug}`;
+
+          return (
+            <tr
+              key={person.slug}
+              data-cy="person"
+              className={classNames({
+                'has-background-warning': isSelected,
+              })}
+            >
+              <td>
+                <PersonLink person={person} />
+              </td>
+
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+
+              <td>
+                {person.motherName ? (
+                  mother ? (
+                    <PersonLink person={mother} />
+                  ) : (
+                    person.motherName
+                  )
+                ) : (
+                  '-'
+                )}
+              </td>
+
+              <td>
+                {person.fatherName ? (
+                  father ? (
+                    <PersonLink person={father} />
+                  ) : (
+                    person.fatherName
+                  )
+                ) : (
+                  '-'
+                )}
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
